requirements.txt
numpy
scipy
matplotlib
pillow
scikit-image
opencv-contrib-python
tqdm